refactor: move VsockUserVsockBackend into a different file
